Refreshed requirements.
Two python modules (GDAL and requests) have to be moved to be
installed via system packages. This is required because GDAL
depends on the proper version of the system libraries being
installed and old versions are no longer available in PyPI. For
requests, the installation of non-system version fails as pip
is also using requests.