tunneldigger: Changed the way aggregates are computed.
Previously, a datastream DynamicSumField was used to compute the
aggregate tx/rx rates using derived streams. Now, we sum the
raw counters in nodewatcher and generate normal CounterField-based
streams.

This approach is much better as it handles the changing interfaces
transparently, while derived streams needed to be dropped and
re-created when the set of interfaces changed.